minor findbugs

+++ core/src/main/java/org/yawni/util/cache/BloomFilter.java
@@ -138,7 +138,7 @@ public final class BloomFilter<E> extends AbstractSet<E> implements Serializable
    * @return         The optimal number of hashing functions.
    */
   private static int numberOfHashes(int capacity, int bits) {
-    return (int) ceil((bits / capacity) * log(2));
+    return (int) ceil((((double)bits) / capacity) * log(2));
   }
 
   /**
